# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "EngineerCMS API",
        "description": "ECMS has every tool to get any job done, so codename for the new ECMS APIs.",
        "version": "1.0.0",
        "contact": {
            "email": "504284@qq.com"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/chat/avatar/{text}": {
            "get": {
                "tags": [
                    "chat"
                ],
                "description": "get avatar",
                "operationId": "ChatController.get avatar",
                "parameters": [
                    {
                        "in": "path",
                        "name": "text",
                        "description": "The text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.Avatar"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/chat/chat": {
            "get": {
                "tags": [
                    "chat"
                ],
                "description": "get chat",
                "operationId": "ChatController.get chat",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.UserTemple"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/chat/wschat": {
            "get": {
                "tags": [
                    "chat"
                ],
                "description": "get wschat",
                "operationId": "ChatController.get wschat",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/.GetChat"
                        }
                    },
                    "400": {
                        "description": "Invalid page supplied"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        ".Avatar": {
            "type": "object"
        },
...
        ".GetChat": {
            "type": "object"
        },
...
        ".UserTemple": {
            "type": "object"
        },
...
    },
...
}